extract the call checking from NativeImagePointsToAnalysis into a dedicated class
