+++ substratevm/src/com.oracle.graal.pointsto/src/com/oracle/graal/pointsto/BigBang.java
@@ -26,6 +26,7 @@ package com.oracle.graal.pointsto;
 
 import com.oracle.graal.pointsto.api.HostVM;
 import com.oracle.graal.pointsto.constraints.UnsupportedFeatures;
+import com.oracle.graal.pointsto.meta.AnalysisMethod;
 import com.oracle.graal.pointsto.meta.AnalysisUniverse;
 import com.oracle.graal.pointsto.meta.HostedProviders;
 import com.oracle.graal.pointsto.util.Timer;
@@ -33,6 +34,7 @@ import jdk.vm.ci.meta.ConstantReflectionProvider;
 import org.graalvm.compiler.api.replacements.SnippetReflectionProvider;
 import org.graalvm.compiler.debug.DebugContext;
 import org.graalvm.compiler.debug.DebugHandlersFactory;
+import org.graalvm.compiler.graph.NodeSourcePosition;
 import org.graalvm.compiler.options.OptionValues;
 
 import java.io.PrintWriter;
@@ -96,4 +98,10 @@ public interface BigBang extends ReachabilityAnalysis, HeapScanning {
     boolean extendedAsserts();
 
     void runAnalysis(DebugContext debug, Function<AnalysisUniverse, Boolean> duringAnalysisAction) throws InterruptedException;
+
+    /** You can blacklist certain callees here. */
+    @SuppressWarnings("unused")
+    default boolean isCallAllowed(PointsToAnalysis bb, AnalysisMethod caller, AnalysisMethod target, NodeSourcePosition srcPosition) {
+        return true;
+    }
 }
